third_party/blink/web_tests/external/wpt/sanitizer-api/sanitizer-names.https-expected.txt

This is a testharness.js-based test.
[PASS] Element names in config item: allowElements
[PASS] Element names in config item: dropElements
[PASS] Element names in config item: blockElements
[PASS] Attribute names in config item: allowAttributes
[PASS] Attribute names in config item: dropAttributes
[FAIL] Namespaced elements #0: allowElements: ["p"]
  assert_equals: expected "<p>Hello</p>" but got ""
[FAIL] Namespaced elements #1: allowElements: ["svg"]
  assert_equals: expected "Hello" but got ""
[FAIL] Namespaced elements #2: allowElements: [{"name":"svg","namespace":"http://www.w3.org/2000/svg"}]
  assert_equals: expected "<svg>Hello</svg>" but got ""
[FAIL] Namespaced elements #3: allowElements: ["math"]
  assert_equals: expected "Hello" but got ""
[FAIL] Namespaced elements #4: allowElements: [{"name":"math","namespace":"http://www.w3.org/2000/svg"}]
  assert_equals: expected "Hello" but got ""
[FAIL] Namespaced elements #5: allowElements: [{"name":"math","namespace":"http://www.w3.org/1998/Math/MathML"}]
  assert_equals: expected "<math>Hello</math>" but got ""
[FAIL] Namespaced attributes #0: allowAttributes: [{"name":"style","elements":"*"}]
  Failed to construct 'Sanitizer': Failed to read the 'allowAttributes' property from 'SanitizerConfig': The object must have a callable @@iterator property.
[FAIL] Namespaced attributes #1: allowAttributes: [{"name":"href","elements":"*"}]
  Failed to construct 'Sanitizer': Failed to read the 'allowAttributes' property from 'SanitizerConfig': The object must have a callable @@iterator property.
[FAIL] Namespaced attributes #2: allowAttributes: [{"name":"xlink:href","elements":"*"}]
  Failed to construct 'Sanitizer': Failed to read the 'allowAttributes' property from 'SanitizerConfig': The object must have a callable @@iterator property.
[FAIL] Namespaced attributes #3: allowAttributes: [{"name":"href","namespace":"http://www.w3.org/1999/xlink","elements":"*"}]
  Failed to construct 'Sanitizer': Failed to read the 'allowAttributes' property from 'SanitizerConfig': The object must have a callable @@iterator property.
[FAIL] Namespaced attributes #4: allowAttributes: [{"name":"href","namespace":"http://www.w3.org/1999/xlink","elements":"*"}]
  Failed to construct 'Sanitizer': Failed to read the 'allowAttributes' property from 'SanitizerConfig': The object must have a callable @@iterator property.
[FAIL] Namespaced attributes #5: allowAttributes: [{"name":"href","elements":"*"}]
  Failed to construct 'Sanitizer': Failed to read the 'allowAttributes' property from 'SanitizerConfig': The object must have a callable @@iterator property.
[FAIL] Namespaced attribute xlink:href inside SVG tree
  Failed to construct 'Sanitizer': Failed to read the 'allowAttributes' property from 'SanitizerConfig': The object must have a callable @@iterator property.
[FAIL] Mixed-case element names #0: "svg:feBlend"
  Cannot read properties of null (reading 'innerHTML')
[FAIL] Lower-case element names #0: "svg:feblend"
  Cannot read properties of null (reading 'innerHTML')
[FAIL] Upper-case element names #0: "svg:FEBLEND"
  Cannot read properties of null (reading 'innerHTML')
[FAIL] Mixed case element names #0: "feBlend" is preserved in config.
  assert_array_equals: lengths differ, expected array ["svg:feBlend"] length 1, got [] length 0
[FAIL] Mixed-case element names #1: "svg:feColorMatrix"
  Cannot read properties of null (reading 'innerHTML')
[FAIL] Lower-case element names #1: "svg:fecolormatrix"
  Cannot read properties of null (reading 'innerHTML')
[FAIL] Upper-case element names #1: "svg:FECOLORMATRIX"
  Cannot read properties of null (reading 'innerHTML')
[FAIL] Mixed case element names #1: "feColorMatrix" is preserved in config.
  assert_array_equals: lengths differ, expected array ["svg:feColorMatrix"] length 1, got [] length 0
[FAIL] Mixed-case element names #2: "svg:textPath"
  Cannot read properties of null (reading 'innerHTML')
[FAIL] Lower-case element names #2: "svg:textpath"
  Cannot read properties of null (reading 'innerHTML')
[FAIL] Upper-case element names #2: "svg:TEXTPATH"
  Cannot read properties of null (reading 'innerHTML')
[FAIL] Mixed case element names #2: "textPath" is preserved in config.
  assert_array_equals: lengths differ, expected array ["svg:textPath"] length 1, got [] length 0
Harness: the test ran to completion.
